CodeAcrossTO business case
In a one day workshop on the topic of Open Data, I was part of a group that studied the problem of and proposed solutions for way finding in courthouses across Ontario. The work was sponsored by Ontario's Ministry of the Attorney General.

After the workshop, I took time to consolidate our progress and results into a supporting document for business cases for the work we feel is needed. Although this specific document is my personal work, it builds on the work a group of about a dozen volunteers over the course of 5-6 hours.