Service Design Jam 2018: Welcome Toronto
Service Design Jam is a recurring event in which volunteers are presented with an open-ended concept, then form teams to use a design thinking framework to identify a problem and propose a solution. All in the space of less than 48 hours.

My team of 4 identified the problem of people who move to Toronto not knowing where to find relevant information to help them find services and rules that are relevant to settling in the city. There is a lot of information available, but how can we connect people who are new to the city to the resources that will help them get a good start once they move here?

Now the problem and a proposed solution have been formulated, I would be happy to work it CivicTech or a similar group to help them refine our idea, build a prototype, and launch a pilot service.